+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 5.4.3 (2021-07-22)
 * Swift Package Manager
   * Adds `NS_EXTENSION_UNAVAILABLE` annotations to methods unavailable for use in app extensions. Fixes (Drop-In issue #343)[https://github.com/braintree/braintree-ios-drop-in/issues/343] for Xcode 13-beta3.
+* ThreeDSecure
+  * Add `cardAddChallenge` to `BTThreeDSecureRequest`
 
 ## 5.4.2 (2021-06-24)
 * Swift Package Manager